Fix iterator test for latest ES6 semantics.
The previous test was just a `for`-`of` syntax test, looping over an object which is not iterable and doing nothing with the result. The new test makes use of the `Symbol.iterator` property to install an iterator on an object, making it iterable, and then tests that using `for`-`of` on that object iterates once, giving the correct value.