Implemented partial test results.
Now, tests in data-es6 can have, in place of a single test function, a number of sub-tests that check specific aspects of the feature. The let/const tests are the first features to be given this treatment.

Test result data in sub-tests is differential: if a browser is missing, the result for the previous version is used instead (with version lineage inferred from comparing the full names of the browsers).

This also redoes some of the CSS along with adding the styles necessary for partial test rows.