Changed the block-scoping tests to check shadowing.
Shadowing (where an inner scope's same named variable shadows an outer-scope identifier, but doesn't replace it) is arguably a more commonplace feature of block-scoping, and the compilers tend to support it even if they don't support identifiers fully leaving scope (as in e.g. temporal dead zone). This closes #305.

The compilers had in fact already taken this into their account for their results, mostly. This corrects Traceur's result for BLFD.