Update io.js 1.0.0 results
On the TC meeting of 07-01-2015, it was approved to upgrade the v8
version of io.js to 3.31.74.1. This will be the version that ships with
the v1.0.0 release and brings long-awaited ES6 features with it,
specifically:

* Binary and octal literals.
* New String methods.
* Block scoping.
* Template literals.

In addition, some `const` and `let` results were also corrected. All
results are based from the latest nightly.